moved screen Home.js from components folder to pages
